Refactor handlePopState
