+++ client/src/containers/explore/index.tsx
@@ -1,5 +1,7 @@
 "use client";
 
+import { useEffect, useRef } from "react";
+
 import { Section as SectionEntity } from "@shared/dto/sections/section.entity";
 
 import { client } from "@/lib/queryClient";
@@ -14,9 +16,38 @@ export default function Explore() {
     { select: (res) => res.body.data },
   );
   const sections = data as SectionEntity[];
+  const ref = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    const handlePopState = () => {
+      const hash = window.location.hash.slice(1);
+      const id = decodeURIComponent(hash || sections[0].slug);
+      const element = document.getElementById(id);
+
+      if (element && ref.current) {
+        const containerRect = ref.current.getBoundingClientRect();
+        const elementRect = element.getBoundingClientRect();
+        const scrollTop =
+          elementRect.top - containerRect.top + ref.current.scrollTop;
+
+        ref.current.scrollTo({
+          top: scrollTop,
+        });
+      }
+    };
+
+    window.addEventListener("popstate", handlePopState);
+    return () => {
+      window.removeEventListener("popstate", handlePopState);
+    };
+  }, [sections]);
 
   return (
-    <div className="space-y-32 overflow-y-auto scroll-smooth pb-32">
+    <div
+      id="sections-container"
+      ref={ref}
+      className="space-y-32 overflow-y-auto scroll-smooth pb-32"
+    >
       {sections.map((s) => (
         <Section
           key={`section-container-${s.slug}`}
