+++ client/src/containers/explore/section/index.tsx
@@ -56,19 +56,6 @@ const Section: React.FC<PropsWithChildren<SectionProps>> = ({
     };
   }, []);
 
-  useEffect(() => {
-    const handlePopState = () => {
-      const hash = window.location.hash.slice(1);
-      const element = document.getElementById(hash || menuItems[0].slug);
-      element?.scrollIntoView({ behavior: "smooth" });
-    };
-
-    window.addEventListener("popstate", handlePopState);
-    return () => {
-      window.removeEventListener("popstate", handlePopState);
-    };
-  }, []);
-
   const handleLinkClick = (sectionId: string) => {
     setPopOverOpen(false);
